feat: improve test coverage for CreateEventModal
