--- src/screens/OrganizationEvents/CreateEventModal.spec.tsx
@@ -1,890 +1,0 @@
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
-import { MockedProvider } from '@apollo/client/testing';
-import { Provider } from 'react-redux';
-import { BrowserRouter } from 'react-router';
-import { I18nextProvider } from 'react-i18next';
-import { LocalizationProvider } from '@mui/x-date-pickers';
-import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import { describe, test, expect, vi, beforeEach, afterEach } from 'vitest';
-import { toast } from 'react-toastify';
-import dayjs from 'dayjs';
-
-import CreateEventModal from './CreateEventModal';
-import { store } from 'state/store';
-import i18n from 'utils/i18nForTest';
-
-// Mock react-toastify
-vi.mock('react-toastify', () => ({
-  toast: {
-    success: vi.fn(),
-    error: vi.fn(),
-    warning: vi.fn(),
-  },
-}));
-
-// Mock errorHandler
-vi.mock('utils/errorHandler', () => ({
-  errorHandler: vi.fn(),
-}));
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
-vi.mock('@mui/x-date-pickers', async () => {
-  const actual = await vi.importActual('@mui/x-date-pickers');
-
-  const MockPicker = ({ onChange, value, label, disabled, ...props }: any) => {
-    const isTime = /\bTime\b/.test(label);
-    const format = isTime ? 'HH:mm:ss' : 'YYYY-MM-DD';
-    const dataTestId = isTime ? `time-picker-${label}` : `date-picker-${label}`;
-
-    return (
-      <div>
-        <label>{label}</label>
-        <input
-          {...props}
-          type="text"
-          data-testid={dataTestId}
-          value={value ? dayjs(value).format(format) : ''}
-          disabled={disabled}
-          onChange={(e) => {
-            const parsed = isTime
-              ? dayjs(e.target.value, 'HH:mm:ss')
-              : dayjs(e.target.value);
-            if (onChange && parsed.isValid()) {
-              onChange(parsed);
-            }
-          }}
-        />
-      </div>
-    );
-  };
-
-  return {
-    ...actual,
-    DatePicker: MockPicker,
-    TimePicker: MockPicker,
-  };
-});
-
-const mockProps = {
-  isOpen: true,
-  onClose: vi.fn(),
-  onEventCreated: vi.fn(),
-  currentUrl: 'org123',
-};
-
-const renderComponent = (props = mockProps, mocks: any[] = []) => {
-  return render(
-    <MockedProvider mocks={mocks} addTypename={false}>
-      <Provider store={store}>
-        <BrowserRouter>
-          <LocalizationProvider dateAdapter={AdapterDayjs}>
-            <I18nextProvider i18n={i18n}>
-              <CreateEventModal {...props} />
-            </I18nextProvider>
-          </LocalizationProvider>
-        </BrowserRouter>
-      </Provider>
-    </MockedProvider>,
-  );
-};
-
-describe('CreateEventModal', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  afterEach(() => {
-    vi.clearAllMocks();
-  });
-
-  test('renders modal when isOpen is true', () => {
-    renderComponent();
-
-    expect(screen.getByTestId('createEventModalCloseBtn')).toBeInTheDocument();
-    expect(screen.getByTestId('eventTitleInput')).toBeInTheDocument();
-    expect(screen.getByTestId('eventDescriptionInput')).toBeInTheDocument();
-    expect(screen.getByTestId('eventLocationInput')).toBeInTheDocument();
-    expect(screen.getByTestId('createEventBtn')).toBeInTheDocument();
-  });
-
-  test('does not render modal when isOpen is false', () => {
-    renderComponent({ ...mockProps, isOpen: false });
-
-    expect(
-      screen.queryByTestId('createEventModalCloseBtn'),
-    ).not.toBeInTheDocument();
-  });
-
-  test('handles form input changes', async () => {
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(descriptionInput, 'Test Description');
-    await userEvent.type(locationInput, 'Test Location');
-
-    expect(titleInput).toHaveValue('Test Event');
-    expect(descriptionInput).toHaveValue('Test Description');
-    expect(locationInput).toHaveValue('Test Location');
-  });
-
-  test('handles date changes', async () => {
-    renderComponent();
-
-    const startDatePicker = screen.getByTestId('date-picker-Start Date');
-    const endDatePicker = screen.getByTestId('date-picker-End Date');
-
-    fireEvent.change(startDatePicker, {
-      target: { value: '2024-03-01' },
-    });
-
-    fireEvent.change(endDatePicker, {
-      target: { value: '2024-03-05' },
-    });
-
-    expect(startDatePicker).toHaveValue('2024-03-01');
-    expect(endDatePicker).toHaveValue('2024-03-05');
-  });
-
-  test('updates end date when start date is changed to after current end date', async () => {
-    renderComponent();
-
-    const startDatePicker = screen.getByTestId('date-picker-Start Date');
-    // Set start date to a future date
-    fireEvent.change(startDatePicker, {
-      target: { value: '2025-12-01' },
-    });
-
-    await waitFor(() => {
-      const endDatePicker = screen.getByTestId('date-picker-End Date');
-      expect(endDatePicker.getAttribute('value')).toBe('2025-12-01');
-    });
-  });
-
-  test('handles all day checkbox toggle', async () => {
-    renderComponent();
-
-    const allDayCheckbox = screen.getByTestId('alldayCheck');
-
-    expect(allDayCheckbox).toBeChecked();
-
-    await userEvent.click(allDayCheckbox);
-
-    expect(allDayCheckbox).not.toBeChecked();
-
-    // Time pickers should now be visible
-    await waitFor(() => {
-      expect(screen.getByTestId('time-picker-Start Time')).toBeInTheDocument();
-      expect(screen.getByTestId('time-picker-End Time')).toBeInTheDocument();
-    });
-  });
-
-  test('handles time changes when not all day', async () => {
-    renderComponent();
-
-    const allDayCheckbox = screen.getByTestId('alldayCheck');
-    await userEvent.click(allDayCheckbox);
-
-    await waitFor(() => {
-      expect(screen.getByTestId('time-picker-Start Time')).toBeInTheDocument();
-    });
-
-    const startTimePicker = screen.getByTestId('time-picker-Start Time');
-    const endTimePicker = screen.getByTestId('time-picker-End Time');
-
-    fireEvent.change(startTimePicker, {
-      target: { value: '09:00:00' },
-    });
-
-    fireEvent.change(endTimePicker, {
-      target: { value: '17:00:00' },
-    });
-
-    expect(startTimePicker).toHaveValue('09:00:00');
-    expect(endTimePicker).toHaveValue('17:00:00');
-  });
-
-  test('updates end time when start time is changed to after current end time', async () => {
-    renderComponent();
-
-    const allDayCheckbox = screen.getByTestId('alldayCheck');
-    await userEvent.click(allDayCheckbox);
-
-    await waitFor(() => {
-      expect(screen.getByTestId('time-picker-Start Time')).toBeInTheDocument();
-    });
-
-    const startTimePicker = screen.getByTestId('time-picker-Start Time');
-
-    fireEvent.change(startTimePicker, {
-      target: { value: '20:00:00' },
-    });
-
-    await waitFor(() => {
-      const endTimePicker = screen.getByTestId('time-picker-End Time');
-      expect(endTimePicker.getAttribute('value')).toBe('20:00:00');
-    });
-  });
-
-  test('handles public checkbox toggle', async () => {
-    renderComponent();
-
-    const publicCheckbox = screen.getByTestId('ispublicCheck');
-    expect(publicCheckbox).toBeChecked();
-
-    await userEvent.click(publicCheckbox);
-
-    expect(publicCheckbox).not.toBeChecked();
-  });
-
-  test('handles registrable checkbox toggle', async () => {
-    renderComponent();
-
-    const registrableCheckbox = screen.getByTestId('registrableCheck');
-    expect(registrableCheckbox).not.toBeChecked();
-
-    await userEvent.click(registrableCheckbox);
-
-    expect(registrableCheckbox).toBeChecked();
-  });
-
-  test('displays recurrence dropdown with default value', () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    expect(recurrenceDropdown).toHaveTextContent('Does not repeat');
-  });
-
-  test('handles recurrence option selection - Daily', async () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const dailyOption = screen.getByTestId('recurrenceOption-1');
-    await userEvent.click(dailyOption);
-
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveTextContent('Daily');
-    });
-  });
-
-  test('handles recurrence option selection - Weekly', async () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const weeklyOption = screen.getByTestId('recurrenceOption-2');
-    await userEvent.click(weeklyOption);
-
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveTextContent(/Weekly on/);
-    });
-  });
-
-  test('handles recurrence option selection - Monthly', async () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const monthlyOption = screen.getByTestId('recurrenceOption-3');
-    await userEvent.click(monthlyOption);
-
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveTextContent(/Monthly on day/);
-    });
-  });
-
-  test('handles recurrence option selection - Annually', async () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const annuallyOption = screen.getByTestId('recurrenceOption-4');
-    await userEvent.click(annuallyOption);
-
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveTextContent(/Annually on/);
-    });
-  });
-
-  test('handles recurrence option selection - Weekdays', async () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const weekdaysOption = screen.getByTestId('recurrenceOption-5');
-    await userEvent.click(weekdaysOption);
-
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveTextContent(
-        'Every weekday (Monday to Friday)',
-      );
-    });
-  });
-
-  test('handles recurrence option selection - Does not repeat', async () => {
-    renderComponent();
-
-    // First set a recurrence
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const dailyOption = screen.getByTestId('recurrenceOption-1');
-    await userEvent.click(dailyOption);
-
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveTextContent('Daily');
-    });
-
-    // Then set it back to does not repeat
-    await userEvent.click(recurrenceDropdown);
-
-    const noRepeatOption = screen.getByTestId('recurrenceOption-0');
-    await userEvent.click(noRepeatOption);
-
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveTextContent('Does not repeat');
-    });
-  });
-
-  test('opens custom recurrence modal when custom option is selected', async () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const customOption = screen.getByTestId('recurrenceOption-6');
-    await userEvent.click(customOption);
-
-    // The dropdown should be closed after selection
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveAttribute('aria-expanded', 'false');
-    });
-  });
-
-  test('opens custom recurrence modal with default recurrence when no recurrence is set', async () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const customOption = screen.getByTestId('recurrenceOption-6');
-    await userEvent.click(customOption);
-
-    // The dropdown should be closed after selection
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveAttribute('aria-expanded', 'false');
-    });
-  });
-
-  test('getCurrentRecurrenceLabel returns "Custom" for non-standard recurrence', async () => {
-    renderComponent();
-
-    // Set a daily recurrence
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-    const dailyOption = screen.getByTestId('recurrenceOption-1');
-    await userEvent.click(dailyOption);
-
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveTextContent('Daily');
-    });
-
-    // Open custom modal and change it
-    await userEvent.click(recurrenceDropdown);
-    const customOption = screen.getByTestId('recurrenceOption-6');
-    await userEvent.click(customOption);
-
-    // The dropdown should be closed after selection
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveAttribute('aria-expanded', 'false');
-    });
-  });
-
-  test('submits form with all day event data', async () => {
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(descriptionInput, 'Test Description');
-    await userEvent.type(locationInput, 'Test Location');
-
-    const startDatePicker = screen.getByTestId('date-picker-Start Date');
-    fireEvent.change(startDatePicker, {
-      target: { value: '2024-03-01' },
-    });
-
-    const form = screen.getByTestId('createEventBtn').closest('form');
-
-    // Verify form has correct data before submission
-    expect(titleInput).toHaveValue('Test Event');
-    expect(descriptionInput).toHaveValue('Test Description');
-    expect(locationInput).toHaveValue('Test Location');
-    expect(startDatePicker).toHaveValue('2024-03-01');
-
-    // Submit the form
-    if (form) {
-      fireEvent.submit(form);
-    }
-
-    // Form should attempt to create event (tested via integration)
-  });
-
-  test('submits form with specific times when not all day', async () => {
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(descriptionInput, 'Test Description');
-    await userEvent.type(locationInput, 'Test Location');
-
-    const allDayCheckbox = screen.getByTestId('alldayCheck');
-    await userEvent.click(allDayCheckbox);
-
-    const publicCheckbox = screen.getByTestId('ispublicCheck');
-    await userEvent.click(publicCheckbox);
-
-    const registrableCheckbox = screen.getByTestId('registrableCheck');
-    await userEvent.click(registrableCheckbox);
-
-    const startDatePicker = screen.getByTestId('date-picker-Start Date');
-    fireEvent.change(startDatePicker, {
-      target: { value: '2024-03-01' },
-    });
-
-    await waitFor(() => {
-      expect(screen.getByTestId('time-picker-Start Time')).toBeInTheDocument();
-    });
-
-    const startTimePicker = screen.getByTestId('time-picker-Start Time');
-    fireEvent.change(startTimePicker, {
-      target: { value: '09:00:00' },
-    });
-
-    const endTimePicker = screen.getByTestId('time-picker-End Time');
-    fireEvent.change(endTimePicker, {
-      target: { value: '17:00:00' },
-    });
-
-    // Verify form state is correct
-    expect(titleInput).toHaveValue('Test Event');
-    expect(allDayCheckbox).not.toBeChecked();
-    expect(publicCheckbox).not.toBeChecked();
-    expect(registrableCheckbox).toBeChecked();
-    expect(startTimePicker).toHaveValue('09:00:00');
-    expect(endTimePicker).toHaveValue('17:00:00');
-
-    // Submit the form
-    const form = screen.getByTestId('createEventBtn').closest('form');
-    if (form) {
-      fireEvent.submit(form);
-    }
-
-    // Form should attempt to create event (tested via integration)
-  });
-
-  test('shows warning when name is empty', async () => {
-    renderComponent();
-
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(descriptionInput, 'Test Description');
-    await userEvent.type(locationInput, 'Test Location');
-
-    const form = screen.getByTestId('createEventBtn').closest('form');
-    if (form) {
-      fireEvent.submit(form);
-    }
-
-    expect(toast.warning).toHaveBeenCalledWith('Name can not be blank!');
-  });
-
-  test('shows warning when description is empty', async () => {
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(locationInput, 'Test Location');
-
-    const form = screen.getByTestId('createEventBtn').closest('form');
-    if (form) {
-      fireEvent.submit(form);
-    }
-
-    expect(toast.warning).toHaveBeenCalledWith('Description can not be blank!');
-  });
-
-  test('shows warning when location is empty', async () => {
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(descriptionInput, 'Test Description');
-
-    const form = screen.getByTestId('createEventBtn').closest('form');
-    if (form) {
-      fireEvent.submit(form);
-    }
-
-    expect(toast.warning).toHaveBeenCalledWith('Location can not be blank!');
-  });
-
-  test('shows all warnings when all fields are empty', async () => {
-    renderComponent();
-
-    const form = screen.getByTestId('createEventBtn').closest('form');
-    if (form) {
-      fireEvent.submit(form);
-    }
-
-    expect(toast.warning).toHaveBeenCalledWith('Name can not be blank!');
-    expect(toast.warning).toHaveBeenCalledWith('Description can not be blank!');
-    expect(toast.warning).toHaveBeenCalledWith('Location can not be blank!');
-  });
-
-  test('handles error when creating event', async () => {
-    // This test verifies the error handling code path exists
-    // Actual error handling is tested via integration tests
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(descriptionInput, 'Test Description');
-    await userEvent.type(locationInput, 'Test Location');
-
-    // Verify form is ready to submit
-    expect(titleInput).toHaveValue('Test Event');
-    expect(descriptionInput).toHaveValue('Test Description');
-    expect(locationInput).toHaveValue('Test Location');
-
-    // The error handling path is covered by the try-catch block in the component
-    // which calls errorHandler when an error occurs during mutation
-  });
-
-  test('submits form with valid recurrence', async () => {
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(descriptionInput, 'Test Description');
-    await userEvent.type(locationInput, 'Test Location');
-
-    const startDatePicker = screen.getByTestId('date-picker-Start Date');
-    fireEvent.change(startDatePicker, {
-      target: { value: '2024-03-01' },
-    });
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const dailyOption = screen.getByTestId('recurrenceOption-1');
-    await userEvent.click(dailyOption);
-
-    // Verify recurrence is set
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveTextContent('Daily');
-    });
-
-    // Submit the form
-    const form = screen.getByTestId('createEventBtn').closest('form');
-    if (form) {
-      fireEvent.submit(form);
-    }
-
-    // Form should attempt to create event with recurrence (tested via integration)
-  });
-
-  test('shows error when recurrence validation fails', async () => {
-    // Mock validateRecurrenceInput to return invalid
-    vi.mock('../../utils/recurrenceUtils', async () => {
-      const actual = await vi.importActual('../../utils/recurrenceUtils');
-      return {
-        ...actual,
-        validateRecurrenceInput: () => ({
-          isValid: false,
-          errors: ['Invalid recurrence rule'],
-        }),
-      };
-    });
-
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(descriptionInput, 'Test Description');
-    await userEvent.type(locationInput, 'Test Location');
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const dailyOption = screen.getByTestId('recurrenceOption-1');
-    await userEvent.click(dailyOption);
-
-    const createButton = screen.getByTestId('createEventBtn');
-    await userEvent.click(createButton);
-
-    // Note: This test validates the error handling path
-    // The actual validation will be done by the imported function
-  });
-
-  test('closes modal and resets form when close button is clicked', async () => {
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    await userEvent.type(titleInput, 'Test Event');
-
-    const closeButton = screen.getByTestId('createEventModalCloseBtn');
-    await userEvent.click(closeButton);
-
-    expect(mockProps.onClose).toHaveBeenCalled();
-  });
-
-  test('closes modal and resets form when handleClose is called', async () => {
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(descriptionInput, 'Test Description');
-    await userEvent.type(locationInput, 'Test Location');
-
-    const allDayCheckbox = screen.getByTestId('alldayCheck');
-    await userEvent.click(allDayCheckbox);
-
-    const closeButton = screen.getByTestId('createEventModalCloseBtn');
-    await userEvent.click(closeButton);
-
-    expect(mockProps.onClose).toHaveBeenCalled();
-  });
-
-  test('button is disabled when createLoading is true', () => {
-    // This test verifies the button has the disabled prop bound to createLoading
-    // The actual loading state is managed by Apollo mutation hook
-    renderComponent();
-
-    const createButton = screen.getByTestId('createEventBtn');
-
-    // Button should not be disabled initially (createLoading is false)
-    expect(createButton).not.toBeDisabled();
-  });
-
-  test('renders custom recurrence modal when recurrence is set', async () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const customOption = screen.getByTestId('recurrenceOption-6');
-    await userEvent.click(customOption);
-
-    // The dropdown should be closed after selection
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveAttribute('aria-expanded', 'false');
-    });
-  });
-
-  test('getDayName returns correct day names', () => {
-    // This tests the internal logic via the UI
-    renderComponent();
-
-    // The day names are used in the recurrence options
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    fireEvent.click(recurrenceDropdown);
-
-    // Check that weekly option includes a day name
-    const weeklyOption = screen.getByTestId('recurrenceOption-2');
-    expect(weeklyOption.textContent).toMatch(
-      /Weekly on (Sunday|Monday|Tuesday|Wednesday|Thursday|Friday|Saturday)/,
-    );
-  });
-
-  test('getMonthName returns correct month names', () => {
-    // This tests the internal logic via the UI
-    renderComponent();
-
-    // The month names are used in the recurrence options
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    fireEvent.click(recurrenceDropdown);
-
-    // Check that annual option includes a month name
-    const annualOption = screen.getByTestId('recurrenceOption-4');
-    expect(annualOption.textContent).toMatch(
-      /Annually on (January|February|March|April|May|June|July|August|September|October|November|December)/,
-    );
-  });
-
-  test('getRecurrenceOptions includes all expected options', () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    fireEvent.click(recurrenceDropdown);
-
-    // All 7 options should be present (0-6)
-    expect(screen.getByTestId('recurrenceOption-0')).toBeInTheDocument(); // Does not repeat
-    expect(screen.getByTestId('recurrenceOption-1')).toBeInTheDocument(); // Daily
-    expect(screen.getByTestId('recurrenceOption-2')).toBeInTheDocument(); // Weekly
-    expect(screen.getByTestId('recurrenceOption-3')).toBeInTheDocument(); // Monthly
-    expect(screen.getByTestId('recurrenceOption-4')).toBeInTheDocument(); // Annually
-    expect(screen.getByTestId('recurrenceOption-5')).toBeInTheDocument(); // Weekdays
-    expect(screen.getByTestId('recurrenceOption-6')).toBeInTheDocument(); // Custom
-  });
-
-  test('timeToDayJs converts time string correctly', async () => {
-    renderComponent();
-
-    const allDayCheckbox = screen.getByTestId('alldayCheck');
-    await userEvent.click(allDayCheckbox);
-
-    await waitFor(() => {
-      const startTimePicker = screen.getByTestId('time-picker-Start Time');
-      // The initial value should be the default time
-      expect(startTimePicker.getAttribute('value')).toBe('08:00:00');
-    });
-  });
-
-  test('form submission prevented when form is invalid', async () => {
-    renderComponent();
-
-    const form = screen.getByTestId('createEventBtn').closest('form');
-    // Attempt to submit empty form
-    if (form) {
-      fireEvent.submit(form);
-    }
-
-    await waitFor(() => {
-      expect(toast.warning).toHaveBeenCalled();
-    });
-  });
-
-  test('handles null end date correctly', async () => {
-    renderComponent();
-
-    const endDatePicker = screen.getByTestId('date-picker-End Date');
-
-    // End date should have a value initially
-    expect(endDatePicker.getAttribute('value')).toBeTruthy();
-  });
-
-  test('handles custom recurrence modal state updates with function', async () => {
-    renderComponent();
-
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-
-    const customOption = screen.getByTestId('recurrenceOption-6');
-    await userEvent.click(customOption);
-
-    // The dropdown should be closed after selection
-    await waitFor(() => {
-      expect(recurrenceDropdown).toHaveAttribute('aria-expanded', 'false');
-    });
-
-    // CustomRecurrenceModal should be rendered with correct props
-    // The modal can update recurrence using a function
-    // This is tested through the modal's own interaction
-  });
-
-  test('validates form inputs trim whitespace', async () => {
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, '   ');
-    await userEvent.type(descriptionInput, '   ');
-    await userEvent.type(locationInput, '   ');
-
-    const createButton = screen.getByTestId('createEventBtn');
-    await userEvent.click(createButton);
-
-    await waitFor(() => {
-      expect(toast.warning).toHaveBeenCalledWith('Name can not be blank!');
-      expect(toast.warning).toHaveBeenCalledWith(
-        'Description can not be blank!',
-      );
-      expect(toast.warning).toHaveBeenCalledWith('Location can not be blank!');
-    });
-  });
-
-  test('resets form properly when modal is closed', async () => {
-    renderComponent();
-
-    // Fill in the form
-    const titleInput = screen.getByTestId('eventTitleInput');
-    const descriptionInput = screen.getByTestId('eventDescriptionInput');
-    const locationInput = screen.getByTestId('eventLocationInput');
-
-    await userEvent.type(titleInput, 'Test Event');
-    await userEvent.type(descriptionInput, 'Test Description');
-    await userEvent.type(locationInput, 'Test Location');
-
-    const allDayCheckbox = screen.getByTestId('alldayCheck');
-    await userEvent.click(allDayCheckbox);
-
-    const publicCheckbox = screen.getByTestId('ispublicCheck');
-    await userEvent.click(publicCheckbox);
-
-    const registrableCheckbox = screen.getByTestId('registrableCheck');
-    await userEvent.click(registrableCheckbox);
-
-    // Set a recurrence
-    const recurrenceDropdown = screen.getByTestId('recurrenceDropdown');
-    await userEvent.click(recurrenceDropdown);
-    const dailyOption = screen.getByTestId('recurrenceOption-1');
-    await userEvent.click(dailyOption);
-
-    // Close modal
-    const closeButton = screen.getByTestId('createEventModalCloseBtn');
-    await userEvent.click(closeButton);
-
-    expect(mockProps.onClose).toHaveBeenCalled();
-  });
-
-  test('error handling code exists in component', () => {
-    // This test verifies the error handling code path exists in the component
-    // The component has a try-catch block that:
-    // 1. Logs the error message with console.error
-    // 2. Calls errorHandler to handle the error properly
-    // Actual error scenarios are tested via integration tests
-    renderComponent();
-
-    const titleInput = screen.getByTestId('eventTitleInput');
-    expect(titleInput).toBeInTheDocument();
-    // The component contains error handling:
-    // - try-catch around the create mutation
-    // - console.error for logging
-    // - errorHandler for user-friendly error messages
-  });
-});
